Price float support and display Seller as seller name

+++ src/components/ProductSubmit.tsx
@@ -2,7 +2,6 @@ import React, { SyntheticEvent, useState } from "react";
 import { Product } from "../Models/Product";
 import { postProductAPI } from "../Services/ProductAPIService";
 export function ProductSubmit() {
-
   const [nameInput, setNameInput] = useState<string>("");
   function nameInputHandler(event: SyntheticEvent) {
     let textBox = event.target as HTMLTextAreaElement;
@@ -25,8 +24,8 @@ export function ProductSubmit() {
     let product: Product = {
       id: 0,
       name: nameInput,
-      price: parseInt(priceInput),
-      seller: parseInt(sellerInput)
+      price: parseFloat(priceInput),
+      seller: parseInt(sellerInput),
     };
     postProductAPI(product);
   }
@@ -35,25 +34,24 @@ export function ProductSubmit() {
     <>
       <h1>Submit a new product</h1>
       <input
-        onChange = {nameInputHandler}
-        value = {nameInput}
+        onChange={nameInputHandler}
+        value={nameInput}
         placeholder="Enter product name"
-        ></input>
-      
+      ></input>
+
       <input
-        onChange = {priceInputHandler}
-        value = {priceInput}
+        onChange={priceInputHandler}
+        value={priceInput}
         placeholder="Enter product price"
-        ></input>
+      ></input>
 
       <input
-        onChange = {sellerInputHandler}
-        value = {sellerInput}
+        onChange={sellerInputHandler}
+        value={sellerInput}
         placeholder="Enter Seller ID"
-        ></input>
+      ></input>
 
-      <button onClick = {buttonClickHandler} >Submit</button>
-
+      <button onClick={buttonClickHandler}>Submit</button>
     </>
   );
 }
